requirements.txt
numpy
pillow
matplotlib

